Allow mediation="optional"
From an implementation perspective, the user agent can always take mediation = "optional" as mediation = "required".

There are two reasons to accept mediation = "optional" too:

(a) first, it is the default value of `mediation`, so it allows an unspecified mediation to fallback to something runnable
(b) second, it allows implementations to take the hint from developers and experiment with ways in which the [user mediation](https://www.w3.org/TR/credential-management-1/#dom-credentialmediationrequirement-optional) can be waived (e.g. on a subsequent presentation).